+++ src/container/srv_container.c
@@ -33,7 +33,7 @@
 
 #include <daos_srv/container.h>
 
-#include <daos_api.h>	/* for daos_prop_alloc/_free() */
+#include <daos_types.h>	/* for daos_prop_alloc/_free() */
 #include <daos/rpc.h>
 #include <daos_srv/pool.h>
 #include <daos_srv/rdb.h>
