+++ src/include/daos/object.h
@@ -25,8 +25,7 @@
 
 #include <daos/common.h>
 #include <daos/tse.h>
-#include <daos_types.h>
-#include <daos_api.h>
+#include <daos.h>
 
 /** object metadata stored in the global OI table of container */
 struct daos_obj_md {
