+++ src/include/daos/task.h
@@ -28,7 +28,7 @@
 #define __DAOS_CLI_TASK_H__
 
 #include <daos_types.h>
-#include <daos_api.h>
+#include <daos.h>
 #include <daos/tse.h>
 #include <daos/common.h>
 #include <daos_task.h>
