+++ src/include/daos_container.h
@@ -1,0 +1,516 @@
+/*
+ * (C) Copyright 2015-2019 Intel Corporation.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * GOVERNMENT LICENSE RIGHTS-OPEN SOURCE SOFTWARE
+ * The Government's rights to use, modify, reproduce, release, perform, display,
+ * or disclose this software are subject to the terms of the Apache License as
+ * provided in Contract No. B609815.
+ * Any reproduction of computer software, computer software documentation, or
+ * portions thereof marked with this legend must also reproduce the markings.
+ */
+/**
+ * \file
+ *
+ * DAOS Container API methods
+ */
+
+#ifndef __DAOS_CONTAINER_H__
+#define __DAOS_CONTAINER_H__
+
+#include <daos_types.h>
+
+#if defined(__cplusplus)
+extern "C" {
+#endif
+
+/**
+ * DAOS_COO_RO opens the container for reading only. This flag conflicts with
+ * DAOS_COO_RW.
+ *
+ * DAOS_COO_RW opens the container for reading and writing. This flag conflicts
+ * with DAOS_COO_RO.
+ *
+ * DAOS_COO_NOSLIP disables the automatic epoch slip at epoch commit time. See
+ * daos_epoch_commit().
+ */
+#define DAOS_COO_RO	(1U << 0)
+#define DAOS_COO_RW	(1U << 1)
+#define DAOS_COO_NOSLIP	(1U << 2)
+
+/** Container information */
+typedef struct {
+	/** Container UUID */
+	uuid_t			ci_uuid;
+	/** Epoch of latest persisten snapshot */
+	daos_epoch_t		ci_lsnapshot;
+	/** Number of snapshots */
+	uint32_t		ci_nsnapshots;
+	/** Epochs of returns snapshots */
+	daos_epoch_t	       *ci_snapshots;
+	/* TODO: add more members, e.g., size, # objects, uid, gid... */
+} daos_cont_info_t;
+
+/** DAOS space reclaim strategy */
+enum {
+	DAOS_RECLAIM_SNAPSHOT,
+	DAOS_RECLAIM_BATCH,
+	DAOS_RECLAIM_TIME,
+};
+
+/** self headling strategy bits */
+#define DAOS_SELF_HEAL_AUTO_EXCLUDE	(1U << 0)
+#define DAOS_SELF_HEAL_AUTO_REBUILD	(1U << 1)
+
+/**
+ * DAOS container property types
+ * valid in rage (DAOS_PROP_CO_MIN, DAOS_PROP_CO_MAX).
+ */
+enum daos_cont_prop_type {
+	DAOS_PROP_CO_MIN = 0x1000,
+	/**
+	 * Label - a string that a user can associated with a container.
+	 * default = ""
+	 */
+	DAOS_PROP_CO_LABEL,
+	/**
+	 * Layout type: unknown, POSIX, MPI-IO, HDF5, Apache Arrow, ...
+	 * default value = DAOS_PROP_CO_LAYOUT_UNKOWN
+	 */
+	DAOS_PROP_CO_LAYOUT_TYPE,
+	/**
+	 * Layout version: specific to middleware for interop.
+	 * default = 1
+	 */
+	DAOS_PROP_CO_LAYOUT_VER,
+	/**
+	 * Checksum on/off + checksum type (CRC16, CRC32, SHA-1 & SHA-2).
+	 * default = DAOS_PROP_CO_CSUM_OFF
+	 */
+	DAOS_PROP_CO_CSUM,
+	/**
+	 * Redundancy factor:
+	 * RF1: no data protection. scratched data.
+	 * RF3: 3-way replication, EC 4+2, 8+2, 16+2
+	 * default = RF1 (DAOS_PROP_CO_REDUN_RF1)
+	 */
+	DAOS_PROP_CO_REDUN_FAC,
+	/**
+	 * Redundancy level: default fault domain level for placement.
+	 * default = rack (DAOS_PROP_CO_REDUN_RACK)
+	 */
+	DAOS_PROP_CO_REDUN_LVL,
+	/**
+	 * Maximum number of snapshots to retain.
+	 */
+	DAOS_PROP_CO_SNAPSHOT_MAX,
+	/** ACL: access control list for container */
+	DAOS_PROP_CO_ACL,
+	/** Compression on/off + compression type */
+	DAOS_PROP_CO_COMPRESS,
+	/** Encryption on/off + encryption type */
+	DAOS_PROP_CO_ENCRYPT,
+	DAOS_PROP_CO_MAX,
+};
+
+/** container layout type */
+enum {
+	DAOS_PROP_CO_LAYOUT_UNKOWN,
+	DAOS_PROP_CO_LAYOUT_POSIX,
+	DAOS_PROP_CO_LAYOUT_MPIIO,
+	DAOS_PROP_CO_LAYOUT_HDF5,
+	DAOS_PROP_CO_LAYOUT_ARROW,
+};
+
+/** container checksum type */
+enum {
+	DAOS_PROP_CO_CSUM_OFF,
+	DAOS_PROP_CO_CSUM_CRC16,
+	DAOS_PROP_CO_CSUM_CRC32,
+	DAOS_PROP_CO_CSUM_SHA1,
+	DAOS_PROP_CO_CSUM_SHA2,
+};
+
+/** container compress type */
+enum {
+	DAOS_PROP_CO_COMPRESS_OFF,
+};
+
+/** container encryption type */
+enum {
+	DAOS_PROP_CO_ENCRYPT_OFF,
+};
+
+/** container redundancy factor */
+enum {
+	DAOS_PROP_CO_REDUN_RF1,
+	DAOS_PROP_CO_REDUN_RF3,
+};
+
+enum {
+	DAOS_PROP_CO_REDUN_RACK,
+	DAOS_PROP_CO_REDUN_NODE,
+};
+
+/**
+ * Create a new container with uuid \a uuid on the storage pool connected
+ * by \a poh.
+ *
+ * \param[in]	poh	Pool connection handle.
+ * \param[in]	uuid	UUID of the new Container.
+ * \param[in]	cont_prop
+ *			Optional, container properties pointer
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ *
+ * \return		These values will be returned by \a ev::ev_error in
+ *			non-blocking mode:
+ *			0		Success
+ *			-DER_INVAL	Invalid parameter
+ *			-DER_NO_HDL	Invalid pool handle
+ *			-DER_NO_PERM	Permission denied
+ *			-DER_UNREACH	network is unreachable
+ */
+int
+daos_cont_create(daos_handle_t poh, const uuid_t uuid, daos_prop_t *cont_prop,
+		 daos_event_t *ev);
+
+/**
+ * Destroy a container identified by \a uuid, all objects within this
+ * container will be destroyed as well.
+ * If there is at least one container opener, and \a force is set to zero, then
+ * the operation completes with DER_BUSY. Otherwise, the container is destroyed
+ * when the operation completes.
+ *
+ * \param[in]	poh	Pool connection handle.
+ * \param[in]	uuid	Container UUID.
+ * \param[in]	force	Container destroy will return failure if the container
+ *			is still busy (outstanding open handles). This parameter
+ *			will force the destroy to proceed even if there is an
+ *			outstanding open handle.
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			Function will run in blocking mode if \a ev is NULL.
+ *
+ * \return		These values will be returned by \a ev::ev_error in
+ *			non-blocking mode:
+ *			0		Success
+ *			-DER_NO_PERM	Permission denied
+ *			-DER_UNREACH	Network is unreachable
+ *			-DER_NONEXIST	Container is nonexistent
+ *			-DER_BUSY	Pool is busy
+ */
+int
+daos_cont_destroy(daos_handle_t poh, const uuid_t uuid, int force,
+		  daos_event_t *ev);
+
+/**
+ * Open an existing container identified by UUID \a uuid. Upon successful
+ * completion, \a coh and \a info, both of which shall be allocated by the
+ * caller, return the container handle and the latest container information
+ * respectively. The resulting container handle has an HCE equal to GHCE, an
+ * LHE equal to DAOS_EPOCH_MAX, and an LRE equal to GHCE.
+ *
+ * \param[in]	poh	Pool connection handle.
+ * \param[in]	uuid	UUID to identify container.
+ * \param[in]	flags	Open mode, represented by the DAOS_COO_ bits.
+ * \param[out]	coh	Returned open handle.
+ * \param[out]	info	Optional, return container information
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ *
+ * \return		These values will be returned by \a ev::ev_error in
+ *			non-blocking mode:
+ *			0		Success
+ *			-DER_INVAL	Invalid parameter
+ *			-DER_UNREACH	Network is unreachable
+ *			-DER_NO_PERM	Permission denied
+ *			-DER_NONEXIST	Container is nonexistent
+ */
+int
+daos_cont_open(daos_handle_t poh, const uuid_t uuid, unsigned int flags,
+	       daos_handle_t *coh, daos_cont_info_t *info, daos_event_t *ev);
+
+/**
+ * Close a container handle. Upon successful completion, the container handle's
+ * epoch hold (i.e., if LHE < DAOS_EPOCH_MAX) is released, and any uncommitted
+ * updates from the container handle are discarded.
+ *
+ * \param[in]	coh	Container open handle.
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ *
+ * \return		These values will be returned by \a ev::ev_error in
+ *			non-blocking mode:
+ *			0		Success
+ *			-DER_UNREACH	Network is unreachable
+ *			-DER_NO_HDL	Invalid container handle
+ */
+int
+daos_cont_close(daos_handle_t coh, daos_event_t *ev);
+
+/**
+ * Convert a local container handle to global representation data which can be
+ * shared with peer processes.
+ * If glob->iov_buf is set to NULL, the actual size of the global handle is
+ * returned through glob->iov_buf_len.
+ * This function does not involve any communication and does not block.
+ *
+ * \param[in]	coh	valid local container handle to be shared
+ * \param[out]	glob	pointer to iov of the buffer to store handle information
+ *
+ * \return		These values will be returned:
+ *			non-blocking mode:
+ *			0		Success
+ *			-DER_INVAL	Invalid parameter
+ *			-DER_NO_HDL	Container handle is nonexistent
+ *			-DER_TRUNC	Buffer in \a glob is too short, larger
+ *					buffer required. In this case the
+ *					required buffer size is returned through
+ *					glob->iov_buf_len.
+ */
+int
+daos_cont_local2global(daos_handle_t coh, daos_iov_t *glob);
+
+/**
+ * Create a local container handle for global representation data.
+ *
+ * \param[in]	poh	Pool connection handle the container belong to
+ * \param[in]	glob	Global (shared) representation of a collective handle
+ *			to be extracted
+ * \param[out]	coh	Returned local container handle
+ *
+ * \return		These values will be returned:
+ *			non-blocking mode:
+ *			0		Success
+ *			-DER_INVAL	Invalid parameter
+ *			-DER_NO_HDL	Pool handle is nonexistent
+ */
+int
+daos_cont_global2local(daos_handle_t poh, daos_iov_t glob, daos_handle_t *coh);
+
+/**
+ * Query container information.
+ *
+ * \param[in]	coh	Container open handle.
+ * \param[out]	info	Returned container information.
+ *			If \a info::ci_snapshots is not NULL, epochs of
+ *			snapshots will be stored in it.
+ *			If \a info::ci_snapshots is NULL, number of snapshots
+ *			will be returned by \a info::ci_nsnapshots.
+ * \param[out]	cont_prop
+ *			Optional, returned container properties
+ *			If it is NULL, then needs not query the properties.
+ *			If cont_prop is non-NULL but its dpp_entries is NULL,
+ *			will query all pool properties, DAOS internally
+ *			allocates the needed buffers and assign pointer to
+ *			dpp_entries.
+ *			If cont_prop's dpp_nr > 0 and dpp_entries is non-NULL,
+ *			will query the properties for specific dpe_type(s), DAOS
+ *			internally allocates the needed buffer for dpe_str or
+ *			dpe_val_ptr, if the dpe_type with immediate value then
+ *			will directly assign it to dpe_val.
+ *			User can free the associated buffer by calling
+ *			daos_prop_free().
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ *
+ * \return		These values will be returned by \a ev::ev_error in
+ *			non-blocking mode:
+ *			0		Success
+ *			-DER_INVAL	Invalid parameter
+ *			-DER_UNREACH	Network is unreachable
+ *			-DER_NO_HDL	Invalid container handle
+ */
+int
+daos_cont_query(daos_handle_t container, daos_cont_info_t *info,
+		daos_prop_t *cont_prop, daos_event_t *ev);
+
+/**
+ * List the names of all user-defined container attributes.
+ *
+ * \param[in]	coh	Container handle.
+ * \param[out]	buffer	Buffer containing concatenation of all attribute
+ *			names, each being null-terminated. No truncation is
+ *			performed and only full names will be returned.
+ *			NULL is permitted in which case only the aggregate
+ *			size will be retrieved.
+ * \param[in,out]
+ *		size	[in]: Buffer size. [out]: Aggregate size of all
+ *			attribute names (excluding terminating null characters),
+ *			regardless of the actual buffer size.
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ */
+int
+daos_cont_list_attr(daos_handle_t coh, char *buffer, size_t *size,
+		    daos_event_t *ev);
+
+/**
+ * Retrieve a list of user-defined container attribute values.
+ *
+ * \param[in]	coh	Container handle
+ * \param[in]	n	Number of attributes
+ * \param[in]	names	Array of \a n null-terminated attribute names.
+ * \param[out]	buffers	Array of \a n buffers to store attribute values.
+ *			Attribute values larger than corresponding buffer sizes
+ *			will be truncated. NULL values are permitted and will be
+ *			treated identical to zero-length buffers, in which case
+ *			only the sizes of attribute values will be retrieved.
+ * \param[in,out]
+ *		sizes	[in]: Array of \a n buffer sizes. [out]: Array of actual
+ *			sizes of \a n attribute values, regardless of given
+ *			buffer sizes.
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ */
+int
+daos_cont_get_attr(daos_handle_t coh, int n, char const *const names[],
+		   void *const buffers[], size_t sizes[], daos_event_t *ev);
+
+/**
+ * Create or update a list of user-defined container attributes.
+ *
+ * \param[in]	coh	Container handle
+ * \param[in]	n	Number of attributes
+ * \param[in]	names	Array of \a n null-terminated attribute names.
+ * \param[in]	values	Array of \a n attribute values
+ * \param[in]	sizes	Array of \a n elements containing the sizes of
+ *			respective attribute values.
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ */
+int
+daos_cont_set_attr(daos_handle_t coh, int n, char const *const names[],
+		   void const *const values[], size_t const sizes[],
+		   daos_event_t *ev);
+
+/**
+ * Allocate a unique set of 64 bit unsigned integers to be used for object ID
+ * generation for that container. This is an optional helper function for
+ * applications to use to guarantee unique object IDs on the container when more
+ * than 1 client are accessing objects on the container. The highest used ID is
+ * tracked in the container metadata for future access to that container. This
+ * doesn't guarantee that the IDs allocated are sequential; and several ID
+ * ranges could be discarded at container close.
+ *
+ * \param[in]	coh	Container open handle.
+ * \param[in]	num_oids
+ *			Number of unique IDs requested.
+ * \param[out]	oid	starting oid that was allocated up to oid + num_oids.
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ *
+ * \return		These values will be returned by \a ev::ev_error in
+ *			non-blocking mode:
+ *			0		Success
+ *			-DER_NO_HDL	Invalid container open handle
+ *			-DER_UNREACH	Network is unreachable
+ */
+int
+daos_cont_alloc_oids(daos_handle_t coh, daos_size_t num_oids, uint64_t *oid,
+		     daos_event_t *ev);
+
+/**
+ * Rollback to a specific persistent snapshot.
+ *
+ * \param[in]	coh	Container handle
+ * \param[in]	epoch	Epoch if persistent snapshot to rollback to.
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ */
+int
+daos_cont_rollback(daos_handle_t coh, daos_epoch_t epoch, daos_event_t *ev);
+
+/**
+ * Subscribe to the container snapshot state. If user specifies a valid epoch,
+ * the call will return once a persistent snapshot has been taken at that epoch
+ * or a greater one. The epoch value will be updated with that epoch. If
+ * multiple snapshots exist at an epoch greater than the one specified, the
+ * lowest one will be returned in the epoch value. If the epoch value passed in
+ * is 0, this call will return the lowest persistent snapshot on the container,
+ * if any exist, otherwise will just wait till a persistent snapshot is created.
+ *
+ * \param[in]	coh	Container handle
+ * \param[in,out]
+ *		epoch	[in]: Epoch of snapshot to wait for. [out]: epoch of
+ *			persistent snapshot that was taken.
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ */
+int
+daos_cont_subscribe(daos_handle_t coh, daos_epoch_t *epoch, daos_event_t *ev);
+
+#define DAOS_SNAPSHOT_MAX_LEN 128
+
+/**
+ * Create a persistent snapshot at the current epoch and return it. The epoch
+ * that is returned can be used to create a read only transaction to read data
+ * from that persistent snapshot. Optionally the snapshot can be given a name as
+ * an attribute which can be retrieved with daos_cont_list_snap(). Name length
+ * can't exceed DAOS_SNAPSHOT_MAX_LEN.
+ *
+ * \param[in]	coh	Container handle
+ * \param[out]	epoch	returned epoch of persistent snapshot taken.
+ * \param[in]	name	Optional null terminated name for snapshot.
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ */
+int
+daos_cont_create_snap(daos_handle_t coh, daos_epoch_t *epoch, char *name,
+		      daos_event_t *ev);
+
+/**
+ * List all the snapshots of a container and optionally retrieve the snapshot
+ * name of each one if it was given at create time.
+ *
+ * \param[in]	coh	Container handle
+ * \param[in,out]
+ *		nr	[in]: Number of snapshots in epochs and names.
+ *			[out]: Actual number of snapshots returned.
+ * \param[out]	epochs	preallocated array of epochs to store snapshots.
+ * \param[out]	names	preallocated array of names of the snapshots.
+ *			DAOS_SNAPSHOT_MAX_LEN can be used for each name
+ *			size if not known.
+ * \param[in,out]
+ *		anchor	Hash anchor for the next call, it should be set to
+ *			zeroes for the first call, it should not be changed
+ *			by caller between calls.
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ */
+int
+daos_cont_list_snap(daos_handle_t coh, int *nr, daos_epoch_t *epochs,
+		    char **names, daos_anchor_t *anchor, daos_event_t *ev);
+
+/**
+ * Destroy a snapshot. The epoch corresponding to the snapshot is not
+ * discarded, but may be aggregated.
+ *
+ * \param[in]	coh	Container handle
+ * \param[in]	epr	Epoch range of snapshots to destroy.
+ *			If epr_lo == epr_hi delete 1 snapshot at epr_lo/hi.
+ *			If epr_lo == 0, delete all snapshots <= epr_hi.
+ *			If epr_hi == DAOS_EPOCH_MAX, delete all snapshots
+ *			>= epr_lo.
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ */
+int
+daos_cont_destroy_snap(daos_handle_t coh, daos_epoch_range_t epr,
+		       daos_event_t *ev);
+
+#if defined(__cplusplus)
+}
+#endif
+#endif /* __DAOS_CONTAINER_H__ */
