+++ src/include/daos_pool.h
@@ -1,0 +1,426 @@
+/*
+ * (C) Copyright 2015-2019 Intel Corporation.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * GOVERNMENT LICENSE RIGHTS-OPEN SOURCE SOFTWARE
+ * The Government's rights to use, modify, reproduce, release, perform, display,
+ * or disclose this software are subject to the terms of the Apache License as
+ * provided in Contract No. B609815.
+ * Any reproduction of computer software, computer software documentation, or
+ * portions thereof marked with this legend must also reproduce the markings.
+ */
+/**
+ * \file
+ *
+ * DAOS Pool API methods
+ */
+
+#ifndef __DAOS_POOL_H__
+#define __DAOS_POOL_H__
+
+#include <daos_types.h>
+
+#if defined(__cplusplus)
+extern "C" {
+#endif
+
+/**
+ * Storage Targets
+ */
+
+/** Type of storage target */
+typedef enum {
+	DAOS_TP_UNKNOWN,
+	/** Rotating disk */
+	DAOS_TP_HDD,
+	/** Flash-based */
+	DAOS_TP_SSD,
+	/** Persistent memory */
+	DAOS_TP_PM,
+	/** Volatile memory */
+	DAOS_TP_VM,
+} daos_target_type_t;
+
+/** Current state of the storage target */
+typedef enum {
+	DAOS_TS_UNKNOWN,
+	/* Not available */
+	DAOS_TS_DOWN_OUT,
+	/* Not available, may need rebuild */
+	DAOS_TS_DOWN,
+	/* Up */
+	DAOS_TS_UP,
+	/* Up and running */
+	DAOS_TS_UP_IN,
+} daos_target_state_t;
+
+/** Description of target performance */
+typedef struct {
+	/** TODO: storage/network bandwidth, latency etc */
+	int			foo;
+} daos_target_perf_t;
+
+enum {
+	DAOS_MEDIA_SCM	= 0,
+	DAOS_MEDIA_NVME,
+	DAOS_MEDIA_MAX
+};
+
+/** Pool target space usage information */
+struct daos_space {
+	/* Total space in bytes */
+	uint64_t	s_total[DAOS_MEDIA_MAX];
+	/* Free space in bytes */
+	uint64_t	s_free[DAOS_MEDIA_MAX];
+};
+
+/** Target information */
+typedef struct {
+	daos_target_type_t	ta_type;
+	daos_target_state_t	ta_state;
+	daos_target_perf_t	ta_perf;
+	struct daos_space	ta_space;
+} daos_target_info_t;
+
+/** Pool space usage information */
+struct daos_pool_space {
+	/* Aggregated space for all live targets */
+	struct daos_space	ps_space;
+	/* Min target free space in bytes */
+	uint64_t	ps_free_min[DAOS_MEDIA_MAX];
+	/* Max target free space in bytes */
+	uint64_t	ps_free_max[DAOS_MEDIA_MAX];
+	/* Average target free space in bytes */
+	uint64_t	ps_free_mean[DAOS_MEDIA_MAX];
+	/* Target(VOS) count */
+	uint32_t	ps_ntargets;
+	uint32_t	ps_padding;
+};
+
+struct daos_rebuild_status {
+	/** pool map version in rebuilding or last completed rebuild */
+	uint32_t		rs_version;
+	/** padding bytes */
+	uint32_t		rs_pad_32;
+	/** errno for rebuild failure */
+	int32_t			rs_errno;
+	/**
+	 * rebuild is done or not, it is valid only if @rs_version is non-zero
+	 */
+	int32_t			rs_done;
+	/** # total to-be-rebuilt objects, it's non-zero and increase when
+	 * rebuilding in progress, when rs_done is 1 it will not change anymore
+	 * and should equal to rs_obj_nr. With both rs_toberb_obj_nr and
+	 * rs_obj_nr the user can know the progress of the rebuilding.
+	 */
+	uint64_t		rs_toberb_obj_nr;
+	/** # rebuilt objects, it's non-zero only if rs_done is 1 */
+	uint64_t		rs_obj_nr;
+	/** # rebuilt records, it's non-zero only if rs_done is 1 */
+	uint64_t		rs_rec_nr;
+};
+
+/**
+ * Storage pool
+ */
+typedef struct {
+	/** Pool UUID */
+	uuid_t				pi_uuid;
+	/** Number of targets */
+	uint32_t			pi_ntargets;
+	/** Number of nodes */
+	uint32_t			pi_nnodes;
+	/** Number of deactivated targets */
+	uint32_t			pi_ndisabled;
+	/** Latest pool map version */
+	uint32_t			pi_map_ver;
+	/** pool UID */
+	uid_t				pi_uid;
+	/** pool GID */
+	gid_t				pi_gid;
+	/** Mode */
+	uint32_t			pi_mode;
+	/** current raft leader */
+	uint32_t			pi_leader;
+	/** Space usage */
+	struct daos_pool_space		pi_space;
+	/** rebuild status */
+	struct daos_rebuild_status	pi_rebuild_st;
+} daos_pool_info_t;
+
+/**
+ * DAOS_PC_RO connects to the pool for reading only.
+ *
+ * DAOS_PC_RW connects to the pool for reading and writing.
+ *
+ * DAOS_PC_EX connects to the pool for reading and writing exclusively. In the
+ * presence of an exclusive pool handle, no connection with DSM_PC_RW is
+ * permitted.
+ *
+ * The three flags above are mutually exclusive.
+ */
+#define DAOS_PC_RO	(1U << 0)
+#define DAOS_PC_RW	(1U << 1)
+#define DAOS_PC_EX	(1U << 2)
+
+#define DAOS_PC_NBITS	3
+#define DAOS_PC_MASK	((1U << DAOS_PC_NBITS) - 1)
+
+/**
+ * DAOS pool property types
+ * valid in range (DAOS_PROP_PO_MIN, DAOS_PROP_PO_MAX)
+ */
+enum daos_pool_prop_type {
+	DAOS_PROP_PO_MIN = 0,
+	/**
+	 * Label - a string that a user can associated with a pool.
+	 * default = ""
+	 */
+	DAOS_PROP_PO_LABEL,
+	/**
+	 * ACL: access control list for pool
+	 * A list of uid/gid that can access the pool RO
+	 * A list of uid/gid that can access the pool RW
+	 * default RW gid = gid of pool create invoker
+	 */
+	DAOS_PROP_PO_ACL,
+	/**
+	 * Reserve space ratio: amount of space to be reserved on each target
+	 * for rebuild purpose. default = 0%.
+	 */
+	DAOS_PROP_PO_SPACE_RB,
+	/**
+	 * Automatic/manual self-healing. default = auto
+	 * auto/manual exclusion
+	 * auto/manual rebuild
+	 */
+	DAOS_PROP_PO_SELF_HEAL,
+	/**
+	 * Space reclaim strategy = time|batched|snapshot. default = snapshot
+	 * time interval
+	 * batched commits
+	 * snapshot creation
+	 */
+	DAOS_PROP_PO_RECLAIM,
+	DAOS_PROP_PO_MAX,
+};
+
+
+/**
+ * Connect to the DAOS pool identified by UUID \a uuid. Upon a successful
+ * completion, \a poh returns the pool handle, and \a info returns the latest
+ * pool information.
+ *
+ * \param[in]	uuid	UUID to identify a pool.
+ * \param[in]	grp	Process set name of the DAOS servers managing the pool
+ * \param[in]	svc	Pool service replica ranks, as reported by
+ *			daos_pool_create().
+ * \param[in]	flags	Connect mode represented by the DAOS_PC_ bits.
+ * \param[out]	poh	Returned open handle.
+ * \param[out]	info	Returned pool info.
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ *
+ * \return		These values will be returned by \a ev::ev_error in
+ *			non-blocking mode:
+ *			0		Success
+ *			-DER_INVAL	Invalid parameter
+ *			-DER_UNREACH	Network is unreachable
+ *			-DER_NO_PERM	Permission denied
+ *			-DER_NONEXIST	Pool is nonexistent
+ */
+int
+daos_pool_connect(const uuid_t uuid, const char *grp,
+		  const d_rank_list_t *svc, unsigned int flags,
+		  daos_handle_t *poh, daos_pool_info_t *info, daos_event_t *ev);
+
+/**
+ * Disconnect from the DAOS pool. It should revoke all the container open
+ * handles of this pool.
+ *
+ * \param[in]	poh	Pool connection handle
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ *
+ * \return		These values will be returned by \a ev::ev_error in
+ *			non-blocking mode:
+ *			0		Success
+ *			-DER_UNREACH	Network is unreachable
+ *			-DER_NO_HDL	Invalid pool handle
+ */
+int
+daos_pool_disconnect(daos_handle_t poh, daos_event_t *ev);
+
+/**
+ * Convert a local pool connection to global representation data which can be
+ * shared with peer processes.
+ * If glob->iov_buf is set to NULL, the actual size of the global handle is
+ * returned through glob->iov_buf_len.
+ * This function does not involve any communication and does not block.
+ *
+ * \param[in]	poh	Valid local pool connection handle to be shared
+ * \param[out]	glob	Pointer to iov of the buffer to store handle information
+ *
+ * \return		These values will be returned:
+ *			non-blocking mode:
+ *			0		Success
+ *			-DER_INVAL	Invalid parameter
+ *			-DER_NO_HDL	Pool  handle is nonexistent
+ *			-DER_TRUNC	Buffer in \a glob is too short, a larger
+ *					buffer is required. In this case the
+ *					required buffer size is returned through
+ *					glob->iov_buf_len.
+ */
+int
+daos_pool_local2global(daos_handle_t poh, daos_iov_t *glob);
+
+/**
+ * Create a local pool connection for global representation data.
+ *
+ * \param[in]	glob	Global (shared) representation of a collective handle
+ *			to be extracted
+ * \param[out]	poh	Returned local pool connection handle
+ *
+ * \return		These values will be returned:
+ *			non-blocking mode:
+ *			0		Success
+ *			-DER_INVAL	Invalid parameter
+ */
+int
+daos_pool_global2local(daos_iov_t glob, daos_handle_t *poh);
+
+/**
+ * Query pool information. User should provide at least one of \a info and
+ * \a tgts as output buffer.
+ *
+ * \param[in]	poh	Pool connection handle.
+ * \param[out]	tgts	Optional, returned storage targets in this pool.
+ * \param[out]	info	Optional, returned pool information.
+ * \param[out]	pool_prop
+ *			Optional, returned pool properties.
+ *			If it is NULL, then needs not query the properties.
+ *			If pool_prop is non-NULL but its dpp_entries is NULL,
+ *			will query all pool properties, DAOS internally
+ *			allocates the needed buffers and assign pointer to
+ *			dpp_entries.
+ *			If pool_prop's dpp_nr > 0 and dpp_entries is non-NULL,
+ *			will query the properties for specific dpe_type(s), DAOS
+ *			internally allocates the needed buffer for dpe_str or
+ *			dpe_val_ptr, if the dpe_type with immediate value then
+ *			will directly assign it to dpe_val.
+ *			User can free the associated buffer by calling
+ *			daos_prop_free().
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ *
+ * \return		These values will be returned by \a ev::ev_error in
+ *			non-blocking mode:
+ *			0		Success
+ *			-DER_INVAL	Invalid parameter
+ *			-DER_UNREACH	Network is unreachable
+ *			-DER_NO_HDL	Invalid pool handle
+ */
+int
+daos_pool_query(daos_handle_t poh, d_rank_list_t *tgts, daos_pool_info_t *info,
+		daos_prop_t *pool_prop, daos_event_t *ev);
+
+/**
+ * Query information of storage targets within a DAOS pool.
+ *
+ * \param[in]	poh	Pool connection handle.
+ * \param[in]	tgts	A list of targets to query.
+ * \param[out]	failed	Optional, buffer to store faulty targets on failure.
+ * \param[out]	info_list
+ *			Returned storage information of \a tgts, it is an array
+ *			and array size must equal to tgts::rl_llen.
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ *
+ * \return		These values will be returned by \a ev::ev_error in
+ *			non-blocking mode:
+ *			0		Success
+ *			-DER_INVAL	Invalid parameter
+ *			-DER_NO_HDL	Invalid pool handle
+ *			-DER_UNREACH	Network is unreachable
+ *			-DER_NONEXIST	No pool on specified targets
+ */
+int
+daos_pool_query_target(daos_handle_t poh, d_rank_list_t *tgts,
+		       d_rank_list_t *failed, daos_target_info_t *info_list,
+		       daos_event_t *ev);
+
+/**
+ * List the names of all user-defined pool attributes.
+ *
+ * \param[in]	poh	Pool handle.
+ * \param[out]	buffer	Buffer containing concatenation of all attribute
+ *			names, each being null-terminated. No truncation is
+ *			performed and only full names will be returned.
+ *			NULL is permitted in which case only the aggregate
+ *			size will be retrieved.
+ * \param[in,out]
+ *		size	[in]: Buffer size. [out]: Aggregate size of all
+ *			attribute names (excluding terminating null
+ *			characters), regardless of the actual buffer
+ *			size.
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ */
+int
+daos_pool_list_attr(daos_handle_t poh, char *buffer, size_t *size,
+		    daos_event_t *ev);
+
+/**
+ * Retrieve a list of user-defined pool attribute values.
+ *
+ * \param[in]	poh	Pool handle
+ * \param[in]	n	Number of attributes
+ * \param[in]	names	Array of \a n null-terminated attribute names.
+ * \param[out]	buffers	Array of \a n buffers to store attribute values.
+ *			Attribute values larger than corresponding buffer sizes
+ *			will be truncated. NULL values are permitted and will be
+ *			treated identical to zero-length buffers, in which case
+ *			only the sizes of attribute values will be retrieved.
+ * \param[in,out]
+ *		sizes	[in]: Array of \a n buffer sizes. [out]: Array of actual
+ *			sizes of \a n attribute values, regardless of given
+ *			buffer sizes.
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ */
+int
+daos_pool_get_attr(daos_handle_t poh, int n, char const *const names[],
+		   void *const buffers[], size_t sizes[], daos_event_t *ev);
+
+/**
+ * Create or update a list of user-defined pool attributes.
+ *
+ * \param[in]	poh	Pool handle
+ * \param[in]	n	Number of attributes
+ * \param[in]	names	Array of \a n null-terminated attribute names.
+ * \param[in]	values	Array of \a n attribute values
+ * \param[in]	sizes	Array of \a n elements containing the sizes of
+ *			respective attribute values.
+ * \param[in]	ev	Completion event, it is optional and can be NULL.
+ *			The function will run in blocking mode if \a ev is NULL.
+ */
+int
+daos_pool_set_attr(daos_handle_t poh, int n, char const *const names[],
+		   void const *const values[], size_t const sizes[],
+		   daos_event_t *ev);
+
+#if defined(__cplusplus)
+}
+#endif
+#endif /* __DAOS_POOL_H__ */
