+++ src/tests/daosctl/cont-cmds.c
@@ -39,8 +39,6 @@
 /* daos specific */
 #include "common_utils.h"
 #include <daos.h>
-#include <daos_api.h>
-#include <daos_mgmt.h>
 #include <daos/common.h>
 
 /**
