+++ src/tests/daosctl/test-pool.c
@@ -38,8 +38,6 @@
 /* daos specific */
 #include "common_utils.h"
 #include <daos.h>
-#include <daos_api.h>
-#include <daos_mgmt.h>
 #include <daos/common.h>
 
 /**
